+++ core/src/components/popover/popover.tsx
@@ -528,7 +528,7 @@ export class Popover implements ComponentInterface, PopoverInterface {
       await waitForMount();
     }
 
-    await present<PopoverPresentOptions>(this, 'popoverEnter', iosEnterAnimation, mdEnterAnimation, {
+    await present<PopoverPresentOptions>(this, 'popoverEnter', iosEnterAnimation, mdEnterAnimation, undefined, {
       event: event || this.event,
       size: this.size,
       trigger: this.triggerEl,