+++ packages/vue/src/proxies.ts
@@ -957,6 +957,7 @@ export const IonSegmentView: StencilVueComponent<JSX.IonSegmentView> = /*@__PURE
 
 export const IonSelect: StencilVueComponent<JSX.IonSelect, JSX.IonSelect["value"]> = /*@__PURE__*/ defineContainer<JSX.IonSelect, JSX.IonSelect["value"]>('ion-select', defineIonSelect, [
   'cancelText',
+  'cancelIcon',
   'color',
   'compareWith',
   'disabled',
@@ -999,6 +1000,7 @@ export const IonSelect: StencilVueComponent<JSX.IonSelect, JSX.IonSelect["value"
 export const IonSelectModal: StencilVueComponent<JSX.IonSelectModal> = /*@__PURE__*/ defineContainer<JSX.IonSelectModal>('ion-select-modal', defineIonSelectModal, [
   'header',
   'cancelText',
+  'cancelIcon',
   'multiple',
   'options'
 ]);
@@ -1006,7 +1008,8 @@ export const IonSelectModal: StencilVueComponent<JSX.IonSelectModal> = /*@__PURE
 
 export const IonSelectOption: StencilVueComponent<JSX.IonSelectOption> = /*@__PURE__*/ defineContainer<JSX.IonSelectOption>('ion-select-option', defineIonSelectOption, [
   'disabled',
-  'value'
+  'value',
+  'description'
 ]);